Consolidated dataset forming scripts
